skip fittmb test if .dll doesn't exist

+++ tests/testthat/test-unit-utils-fittmb.R
@@ -9,6 +9,7 @@ context("unit test utils-fit.tmb")
 
 test_that("run.model = FALSE", {
   skip_on_cran()
+  skip_if_not(run_model)
   clustTMB.mod <- fit.tmb(args,
     control = list(run.model = FALSE)
   )
@@ -28,6 +29,7 @@ test_that("run.model = FALSE", {
 
 test_that("run.model = TRUE, sdr.report = FALSE", {
   skip_on_cran()
+  skip_if_not(run_model)
   clustTMB.mod <- fit.tmb(args,
     control = list(
       run.model = TRUE,
@@ -59,6 +61,7 @@ test_that("run.model = TRUE, sdr.report = FALSE", {
 
 test_that("run.model = TRUE, sdr.report = TRUE", {
   skip_on_cran()
+  skip_if_not(run_model)
   clustTMB.mod <- fit.tmb(args,
     control = list(
       run.model = TRUE,
@@ -89,6 +92,8 @@ test_that("run.model = TRUE, sdr.report = TRUE", {
   expect_equal(sdr, clustTMB.mod$sdr)
 })
 
-dyn.unload(dynlib(testthat::test_path("ref", "simple")))
-file.remove(dynlib(testthat::test_path("ref", "simple")))
-file.remove(testthat::test_path("ref", "simple.o"))
+if(run_model){
+  dyn.unload(dynlib(testthat::test_path("ref", "simple")))
+  file.remove(dynlib(testthat::test_path("ref", "simple")))
+  file.remove(testthat::test_path("ref", "simple.o"))
+}
